# Conflicts:
#	src/service/ProductService.ts

+++ src/controller/ProductController.ts
@@ -1,4 +1,3 @@
-
 import categoryService from "../service/CategoryService";
 import { Request, Response } from "express";
 import productService from "../service/ProductService";
@@ -45,14 +44,17 @@ class ProductController {
    editProduct = async (req: Request, res: Response) => {
       try {
          let idProduct = req.params.id;
-         // let idUser = req["decoded"].idUser;
-         // let check = await this.productService.checkUser(idUser, idProduct);
-         // if (check) {
-         let products = await this.productService.update(idProduct, req.body);
-         return res.status(200).json(products);
-         // } else {
-         //   return res.status(401).json("invalid");
-         // }
+         let idUser = req["decoded"].idUser;
+         let check = await this.productService.checkUser(idUser, idProduct);
+         if (check) {
+            let products = await this.productService.update(
+               idProduct,
+               req.body
+            );
+            return res.status(200).json(products);
+         } else {
+            return res.status(401).json("invalid");
+         }
       } catch (e) {
          res.status(500).json(e.message);
       }
@@ -60,14 +62,14 @@ class ProductController {
    deleteProduct = async (req: Request, res: Response) => {
       try {
          let idProduct = req.params.id;
-         // let idUser = req["decoded"].idUser;
-         // let check = await this.productService.checkUser(idUser, idProduct);
-         // if (check || req["decoded"].role === "admin") {
-         let products = await this.productService.delete(idProduct);
-         return res.status(200).json(products);
-         // } else {
-         //   return res.status(401).json("invalid");
-         // }
+         let idUser = req["decoded"].idUser;
+         let check = await this.productService.checkUser(idUser, idProduct);
+         if (check || req["decoded"].role === "admin") {
+            let products = await this.productService.delete(idProduct);
+            return res.status(200).json(products);
+         } else {
+            return res.status(401).json("invalid");
+         }
       } catch (e) {
          res.status(500).json(e.message);
       }
@@ -89,26 +91,58 @@ class ProductController {
          res.status(500).json(e.message);
       }
    };
-    getAll = async (req: Request, res: Response) => {
-        try {
-            let product = await productService.getAll()
-            res.status(200).json(product)
-        } catch (e) {
-            res.status(500).json(e.message)
-        }
-
-
-    }
-    search = async (req: Request, res: Response) => {
-        try {
-            let product = await productService.search(req,res)
-            res.status(200).json(product)
-
-        }catch (e) {
-            res.status(500).json(e.message)
-        }
-
-    }
+   getAll = async (req: Request, res: Response) => {
+      try {
+         let product = await productService.getAll();
+         res.status(200).json(product);
+      } catch (e) {
+         res.status(500).json(e.message);
+      }
+   };
+   search = async (req: Request, res: Response) => {
+      try {
+         let limit = 20;
+         let offset = 0;
+         let page = 1;
+         if (req.query.page) {
+            page = +req.query.page;
+            offset = (+page - 1) * limit;
+         }
+         let products = await productService.search(req, res, limit, offset);
+         return res.status(201).json({
+            products: products.products,
+            currentPage: page,
+            totalPage: products.totalPage,
+         });
+      } catch (e) {
+         res.status(500).json(e.message);
+      }
+   };
+   findProductByIdShop = async (req: Request, res: Response) => {
+      try {
+         let limit = 30;
+         let offset = 0;
+         let page = 1;
+         if (req.query.page) {
+            page = +req.query.page;
+            offset = (+page - 1) * limit;
+         }
+         let idProduct = req.params.id;
+         let products = await productService.findByIdShop(
+            idProduct,
+            limit,
+            offset
+         );
+         return res.status(201).json({
+            products: products.products,
+            currentPage: page,
+            totalPage: products.totalPage,
+            count: products.count
+         });
+      } catch (e) {
+         res.status(500).json(e.message);
+      }
+   };
 }
 
 export default new ProductController();
